+++ server/src/foldingRanges.ts
@@ -17,6 +17,8 @@
 import { FoldingRange, FoldingRangeKind } from "vscode-languageserver";
 import { DtcBaseNode } from "./ast/dtc/node";
 import { Parser } from "./parser";
+import { ASTBase } from "./ast/base";
+import { IfDefineBlock } from "./ast/cPreprocessors/ifDefine";
 
 const nodeToRange = (dtcNode: DtcBaseNode): FoldingRange[] => {
   if (!dtcNode.openScope || !dtcNode.closeScope?.prevToken) {
@@ -32,8 +34,41 @@ const nodeToRange = (dtcNode: DtcBaseNode): FoldingRange[] => {
 
   return [range, ...dtcNode.nodes.flatMap(nodeToRange)];
 };
+
+const ifDefBlockToRange = (ifDefBlock: IfDefineBlock): FoldingRange[] => {
+  const ranges: FoldingRange[] = [];
+
+  [ifDefBlock.ifDef.content, ifDefBlock.elseOption?.content].forEach((b) => {
+    if (!b) {
+      return;
+    }
+
+    ranges.push({
+      startLine: b.firstToken.prevToken!.pos.line,
+      startCharacter: b.firstToken.prevToken!.pos.col,
+      endLine: b.lastToken.pos.line,
+      endCharacter: b.lastToken.pos.col + b.lastToken.pos.len,
+      kind: FoldingRangeKind.Region,
+    });
+  });
+
+  return ranges;
+};
+
+const toFoldingRange = (ast: ASTBase): FoldingRange[] => {
+  if (ast instanceof DtcBaseNode) {
+    return nodeToRange(ast);
+  }
+
+  if (ast instanceof IfDefineBlock) {
+    return ifDefBlockToRange(ast);
+  }
+
+  return [];
+};
+
 export function getFoldingRanges(uri: string, parser: Parser): FoldingRange[] {
-  return parser.rootDocument.nodes
+  return parser.allAstItems
     .filter((n) => n.uri === uri)
-    .flatMap(nodeToRange);
+    .flatMap(toFoldingRange);
 }